Use type uint8 to store binary weights.

+++ cpp/subprojects/common/src/common/sampling/weight_sampling.hpp
@@ -27,9 +27,8 @@ static inline std::unique_ptr<IWeightVector> sampleWeightsWithoutReplacementViaT
                                                                                                  uint32 numSamples,
                                                                                                  uint32 numElements,
                                                                                                  RNG& rng) {
-    std::unique_ptr<DenseWeightVector<uint32>> weightVectorPtr = std::make_unique<DenseWeightVector<uint32>>(
-        numElements);
-    typename DenseWeightVector<uint32>::iterator sampleIterator = weightVectorPtr->begin();
+    std::unique_ptr<DenseWeightVector<uint8>> weightVectorPtr = std::make_unique<DenseWeightVector<uint8>>(numElements);
+    typename DenseWeightVector<uint8>::iterator sampleIterator = weightVectorPtr->begin();
     std::unordered_set<uint32> selectedIndices;
 
     for (uint32 i = 0; i < numSamples; i++) {
@@ -66,9 +65,8 @@ template<class T>
 static inline std::unique_ptr<IWeightVector> sampleWeightsWithoutReplacementViaPool(T iterator, uint32 numTotal,
                                                                                     uint32 numSamples,
                                                                                     uint32 numElements, RNG& rng) {
-    std::unique_ptr<DenseWeightVector<uint32>> weightVectorPtr = std::make_unique<DenseWeightVector<uint32>>(
-        numElements);
-    typename DenseWeightVector<uint32>::iterator sampleIterator = weightVectorPtr->begin();
+    std::unique_ptr<DenseWeightVector<uint8>> weightVectorPtr = std::make_unique<DenseWeightVector<uint8>>(numElements);
+    typename DenseWeightVector<uint8>::iterator sampleIterator = weightVectorPtr->begin();
     uint32 pool[numTotal];
 
     // Initialize pool...
